if (score >= 1 && score <= 10) { return { redFlags: [ "A faint whisper of awkwardness. Barely detectable.", "So subtle, the cringe almost slipped by unnoticed.", "You're flying low on the cringe radar.", ], explanation: "Very low cringe. Minor quirks at best, nothing to raise an alarm.", };
the code displays what you did, you just added the conditions for 1% to 10%, but i asked you to creates roasts for individual percentage like for 1% it will be a brutal roast and for 2% it will be more brutal than 1% and so on till 100%, and also if you will use switch or if else statements then the roasts should not be like this:
"Score 98: Cringe-o-meter is baffled. You're off the charts!". A single line roasts does not fit well with the truthcard, so add atleast 3 roasts statements in which ever condition statements you are using.

that was my idea so that when user uploads they do not get repeated roasts.

implement the above.